Revert "Unit tests for insolvency date page"
This reverts commit e9866c93011ec4a709ea521e6125c9eedfa87682.
