+++ Frontend/CO.CDP.OrganisationApp/Pages/Supplier/ConnectedEntity/ConnectedEntityCompanyInsolvencyDate.cshtml
@@ -9,8 +9,7 @@
     var hasError = dateOfInsolvencyHasError || dayHasError || monthHasError || yearHasError;
 }
 
-<a href="/organisation/@Model.Id/supplier-information/connected-person/company-question/@Model.ConnectedEntityId" class="govuk-back-link">Back</a>
-
+<a href="/organisation/@Model.Id/supplier-information/connected-person/nature-of-control/@Model.ConnectedEntityId" class="govuk-back-link">Back</a>
 
 <main class="govuk-main-wrapper">
     <div class="govuk-grid-row">
